[#15] Rename to SurveyRepository and minor naming refactors
