+++ lib/api/repository/home_repository.dart
@@ -5,24 +5,24 @@ import 'package:survey_flutter/api/storage/shared_preference.dart';
 import 'package:survey_flutter/di/provider/dio_provider.dart';
 import 'package:survey_flutter/model/response/survey_data_response.dart';
 
-final homeRepositoryProvider = Provider<HomeRepository>((ref) {
+final surveyRepositoryProvider = Provider<SurveyRepository>((ref) {
   final sharedPreference = ref.watch(sharedPreferenceProvider);
-  return HomeRepositoryImpl(
+  return SurveyRepositoryImpl(
     ApiService(DioProvider().getDioAuthorized(sharedPreference)),
   );
 });
 
-abstract class HomeRepository {
+abstract class SurveyRepository {
   Future<SurveyDataResponse> getSurveys({
     required int pageNumber,
     required int pageSize,
   });
 }
 
-class HomeRepositoryImpl extends HomeRepository {
+class SurveyRepositoryImpl extends SurveyRepository {
   final ApiService _apiService;
 
-  HomeRepositoryImpl(this._apiService);
+  SurveyRepositoryImpl(this._apiService);
 
   @override
   Future<SurveyDataResponse> getSurveys(
